Handle Task HealthChange without TaskID.
Marathon moved from tasks to instances to support pods.
We didn't want to add pods support right now but we need
to handle events that are not backward compatible. To do this
I introduced `TaskID()` that returns task id from instanceId if
task id was not present in event. This is shortcut just to enable
marathon-consul to be used with Marathon 1.4. We need to rethink
if/how we want to support pods and changed Marathon API.