Recheck if service should be deregistered (#190)
When service is marked to be deregisterd during sync ask marathon for
tasks to recheck if this task was not started after sync procedure started.
Prevents deregistartion of tasks started after sync asked for all
marathon services. It also minimize call to marathon for a price
of not deregistering tasks that was killed during sync. This will
be handled by events or next sync.

Fixes: #88